+++ bitswap/client/internal/providerquerymanager/providerquerymanager.go
@@ -126,7 +126,6 @@ func (pqm *ProviderQueryManager) FindProvidersAsync(sessionCtx context.Context, 
 
 	var span trace.Span
 	sessionCtx, span = internal.StartSpan(sessionCtx, "ProviderQueryManager.FindProvidersAsync", trace.WithAttributes(attribute.Stringer("cid", k)))
-	defer span.End()
 
 	select {
 	case pqm.providerQueryMessages <- &newProvideQueryMessage{
@@ -137,6 +136,7 @@ func (pqm *ProviderQueryManager) FindProvidersAsync(sessionCtx context.Context, 
 	case <-pqm.ctx.Done():
 		ch := make(chan peer.ID)
 		close(ch)
+		span.End()
 		return ch
 	case <-sessionCtx.Done():
 		ch := make(chan peer.ID)
@@ -152,14 +152,15 @@ func (pqm *ProviderQueryManager) FindProvidersAsync(sessionCtx context.Context, 
 	case <-pqm.ctx.Done():
 		ch := make(chan peer.ID)
 		close(ch)
+		span.End()
 		return ch
 	case receivedInProgressRequest = <-inProgressRequestChan:
 	}
 
-	return pqm.receiveProviders(sessionCtx, k, receivedInProgressRequest)
+	return pqm.receiveProviders(sessionCtx, k, receivedInProgressRequest, func() { span.End() })
 }
 
-func (pqm *ProviderQueryManager) receiveProviders(sessionCtx context.Context, k cid.Cid, receivedInProgressRequest inProgressRequest) <-chan peer.ID {
+func (pqm *ProviderQueryManager) receiveProviders(sessionCtx context.Context, k cid.Cid, receivedInProgressRequest inProgressRequest, onCloseFn func()) <-chan peer.ID {
 	// maintains an unbuffered queue for incoming providers for given request for a given session
 	// essentially, as a provider comes in, for a given CID, we want to immediately broadcast to all
 	// sessions that queried that CID, without worrying about whether the client code is actually
@@ -171,6 +172,7 @@ func (pqm *ProviderQueryManager) receiveProviders(sessionCtx context.Context, k 
 
 	go func() {
 		defer close(returnedProviders)
+		defer onCloseFn()
 		outgoingProviders := func() chan<- peer.ID {
 			if len(receivedProviders) == 0 {
 				return nil
@@ -261,7 +263,7 @@ func (pqm *ProviderQueryManager) findProviderWorker() {
 					span.AddEvent("ConnectedToProvider", trace.WithAttributes(attribute.Stringer("peer", p)))
 					select {
 					case pqm.providerQueryMessages <- &receivedProviderMessage{
-						ctx: findProviderCtx,
+						ctx: fpr.ctx,
 						k:   k,
 						p:   p,
 					}:
@@ -274,7 +276,7 @@ func (pqm *ProviderQueryManager) findProviderWorker() {
 			cancel()
 			select {
 			case pqm.providerQueryMessages <- &finishedProviderQueryMessage{
-				ctx: findProviderCtx,
+				ctx: fpr.ctx,
 				k:   k,
 			}:
 			case <-pqm.ctx.Done():
@@ -355,7 +357,7 @@ func (rpm *receivedProviderMessage) debugMessage() {
 func (rpm *receivedProviderMessage) handle(pqm *ProviderQueryManager) {
 	requestStatus, ok := pqm.inProgressRequestStatuses[rpm.k]
 	if !ok {
-		log.Errorf("Received provider (%s) for cid (%s) not requested", rpm.p.String(), rpm.k.String())
+		log.Debugf("Received provider (%s) for cid (%s) not requested", rpm.p.String(), rpm.k.String())
 		return
 	}
 	requestStatus.providersSoFar = append(requestStatus.providersSoFar, rpm.p)